refactor: separate stores
