+++ apps/just-func/play-react/src/App.tsx
@@ -2,7 +2,8 @@ import { lazyImport, useStore } from '@just-web/react'
 import { Suspense } from 'react'
 import MainBackdrop from './components/MainBackdrop/MainBackdrop'
 import MainContent from './components/MainContent/MainContent'
-import { getStore, getStoreValue } from './store'
+import { store } from './docViews/store'
+import { getStoreValue } from './store'
 
 const CommandPalette = lazyImport(
   () => getStoreValue().app,
@@ -10,12 +11,11 @@ const CommandPalette = lazyImport(
   m => m.CommandPalette
 )
 function App() {
-  const store = getStore()
-  const [hasDoc] = useStore(store, s => s.docs.length !== 0)
+  const [hasView] = useStore(store, s => s.views.length !== 0)
 
   return (<>
     <main className="bg-zinc-800 min-h-screen text-white">
-      {hasDoc ? <MainContent /> : <MainBackdrop />}
+      {hasView ? <MainContent /> : <MainBackdrop />}
     </main>
     <Suspense fallback={<div>Loading...</div>}>
       <CommandPalette />
