+++ apps/just-func/play-react/src/docViews/store.ts
@@ -1,0 +1,4 @@
+import { createStore } from '@just-web/app'
+import { View } from './types'
+
+export const store = createStore<{ views: View[] }>({ views: [] })
