+++ apps/just-func/play-react/src/docViews/types.ts
@@ -1,0 +1,10 @@
+import { Doc } from '../docs/types'
+
+export interface DocView {
+  type: 'doc',
+  id: string,
+  doc: Doc
+}
+
+export type View = DocView
+
