+++ apps/just-func/play-react/src/docs/store.ts
@@ -1,0 +1,4 @@
+import { createStore } from '@just-web/app'
+import { Doc } from './types'
+
+export const store = createStore<{ docs: Doc[] }>({ docs: [] })
