+++ apps/just-func/play-react/src/docs/types.ts
@@ -1,0 +1,5 @@
+
+export interface Doc {
+  name: string,
+  content: string
+}
